# swagger.yaml
openapi: 3.0.0
info:
  title: Ride Sharing API
  description: API for managing ride-sharing services
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/RS
    description: Local server
paths:
...
  /driver/createDriver:
    post:
      summary: Add a new driver
      tags:
        - Driver
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverRequest'
      responses:
        '201':
          description: New driver added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '403':
          description: User is not authorized as a driver
        '404':
          description: User not found
        '500':
          description: Server error
  /driver/getInfoDriver:
    get:
      summary: Retrieve information about all drivers
      tags:
        - Driver
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of drivers retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Driver'
        '500':
          description: Server error
  /driver/myDriverInfo:
    get:
      summary: Get the driver profile associated with the current user
      tags:
        - Driver
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Driver profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '404':
          description: Driver profile not found for this user
        '500':
          description: Server error
  /driver/updateMyDriverInfo:
    put:
      summary: Update the driver information for the current user
      tags:
        - Driver
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverRequest'
      responses:
        '200':
          description: Driver information updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '404':
          description: Driver profile not found for this user
        '500':
          description: Server error 
  /passenger/createPassenger:
    post:
      summary: Add a new passenger
      tags:
        - Passenger
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PassengerRequest'
      responses:
        '201':
          description: New passenger added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Passenger'
        '403':
          description: User is not authorized as a passenger
        '404':
          description: User not found
        '500':
          description: Server error
  /passenger/getInfoPassenger:
    get:
      summary: Retrieve information about all passengers
      tags:
        - Passenger
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of passengers retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Passenger'
        '500':
          description: Server error
  /passenger/myPassengerInfo:
    get:
      summary: Get the passenger profile associated with the current user
      tags:
        - Passenger
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Passenger profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Passenger'
        '404':
          description: Passenger profile not found for this user
        '500':
          description: Server error
  /passenger/updateMyPassengerInfo:
    put:
      summary: Update the passenger information for the current user
      tags:
        - Passenger
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PassengerRequest'
      responses:
        '200':
          description: Passenger information updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Passenger'
        '404':
          description: Passenger profile not found for this user
        '500':
          description: Server error             
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []